Move apply_ffmpeg to its own module
