+++ main_data_processor.py
@@ -1,7 +1,5 @@
 import logging
-import subprocess
 from typing import Tuple, Optional
-import time
 from dane.config import cfg
 from dane.s3_util import validate_s3_uri
 from io_util import (
@@ -28,6 +26,7 @@ from dane.provenance import (
     generate_initial_provenance,
     stop_timer_and_persist_provenance_chain,
 )
+from apply_ffmpeg import apply_ffmpeg
 
 
 logger = logging.getLogger(__name__)
@@ -113,47 +112,6 @@ def run(input_file_path: str) -> Tuple[CallbackResponse, Optional[Provenance]]:
     logger.info("Results after applying desired I/O")
     logger.info(validated_output)
     return validated_output, full_provenance_chain
-
-
-def apply_ffmpeg(
-    video_input: AudioExtractionInput,
-) -> AudioExtractionOutput:
-    logger.info("Starting model application")
-    start = time.time() * 1000  # convert to ms
-    destination = get_output_file_path(video_input.source_id, OutputType.AUDIO)
-    ffmpeg_cmd = ["ffmpeg", "-i", video_input.input_file_path]
-    if cfg.AUDIO_EXTRACTION_SETTINGS.CONVERT_TO_MONO:
-        ffmpeg_cmd = ffmpeg_cmd + ["-ac", "1"]
-    if cfg.AUDIO_EXTRACTION_SETTINGS.OUTPUT_SAMPLERATE_HZ != 0:
-        ffmpeg_cmd = ffmpeg_cmd + [
-            "-ar",
-            str(cfg.AUDIO_EXTRACTION_SETTINGS.OUTPUT_SAMPLERATE_HZ),
-        ]
-    subprocess.call(ffmpeg_cmd + [destination])
-    logger.info("Executed command:")
-    logger.info(ffmpeg_cmd + [destination])
-    end = time.time() * 1000  # convert to ms
-
-    model_application_provenance = Provenance(
-        activity_name="video2audio",
-        activity_description="converted a video file to audio format",
-        input_data=video_input.input_file_path,
-        start_time_unix=start,
-        parameters=cfg.AUDIO_EXTRACTION_SETTINGS,
-        software_version="0.1.0",
-        output_data=destination,
-        processing_time_ms=end - start,
-    )
-
-    if not model_application_provenance:
-        return AudioExtractionOutput(500, "Failed to apply model")
-
-    return AudioExtractionOutput(
-        200,
-        "Succesfully applied model",
-        get_base_output_dir(video_input.source_id),
-        model_application_provenance,
-    )
 
 
 # assesses the output and makes sure input & output is handled properly
